// supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
// ...
  public: {
    Tables: {
// ...
      attendees: {
        Row: {
          attendeeid: string
          attendeetype: Database["public"]["Enums"]["attendee_type"]
          contactpreference: Database["public"]["Enums"]["attendee_contact_preference"]
          createdat: string
          dietaryrequirements: string | null
          eventtitle: string | null
          person_id: string | null
          registrationid: string
          relatedattendeeid: string | null
          relationship: string | null
          specialneeds: string | null
          updatedat: string
        }
        Insert: {
          attendeeid?: string
          attendeetype: Database["public"]["Enums"]["attendee_type"]
          contactpreference: Database["public"]["Enums"]["attendee_contact_preference"]
          createdat?: string
          dietaryrequirements?: string | null
          eventtitle?: string | null
          person_id?: string | null
          registrationid: string
          relatedattendeeid?: string | null
          relationship?: string | null
          specialneeds?: string | null
          updatedat?: string
        }
        Update: {
          attendeeid?: string
          attendeetype?: Database["public"]["Enums"]["attendee_type"]
          contactpreference?: Database["public"]["Enums"]["attendee_contact_preference"]
          createdat?: string
          dietaryrequirements?: string | null
          eventtitle?: string | null
          person_id?: string | null
          registrationid?: string
          relatedattendeeid?: string | null
          relationship?: string | null
          specialneeds?: string | null
          updatedat?: string
        }
        Relationships: [
          {
            foreignKeyName: "attendees_person_id_fkey"
            columns: ["person_id"]
            isOneToOne: false
            referencedRelation: "people"
            referencedColumns: ["person_id"]
          },
          {
            foreignKeyName: "attendees_registrationid_fkey"
            columns: ["registrationid"]
            isOneToOne: false
            referencedRelation: "registration_payments"
            referencedColumns: ["registration_id"]
          },
          {
            foreignKeyName: "attendees_registrationid_fkey"
            columns: ["registrationid"]
            isOneToOne: false
            referencedRelation: "registration_summary"
            referencedColumns: ["registration_id"]
          },
          {
            foreignKeyName: "attendees_registrationid_fkey"
            columns: ["registrationid"]
            isOneToOne: false
            referencedRelation: "Registrations"
            referencedColumns: ["registration_id"]
          },
          {
            foreignKeyName: "attendees_relatedattendeeid_fkey"
            columns: ["relatedattendeeid"]
            isOneToOne: false
            referencedRelation: "attendees"
            referencedColumns: ["attendeeid"]
          },
        ]
      }
// ...
      registrations: {
        Row: {
          agree_to_terms: boolean | null
          created_at: string | null
          customer_id: string | null
          event_id: string | null
          payment_status: Database["public"]["Enums"]["payment_status"] | null
          primary_attendee_id: string | null
          registration_data: Json[] | null
          registration_date: string | null
          registration_id: string
          registration_type:
            | Database["public"]["Enums"]["registration_type"]
            | null
          status: string | null
          stripe_payment_intent_id: string | null
          total_amount_paid: number | null
          total_price_paid: number | null
          updated_at: string | null
        }
        Insert: {
          agree_to_terms?: boolean | null
          created_at?: string | null
          customer_id?: string | null
          event_id?: string | null
          payment_status?: Database["public"]["Enums"]["payment_status"] | null
          primary_attendee_id?: string | null
          registration_data?: Json[] | null
          registration_date?: string | null
          registration_id: string
          registration_type?:
            | Database["public"]["Enums"]["registration_type"]
            | null
          status?: string | null
          stripe_payment_intent_id?: string | null
          total_amount_paid?: number | null
          total_price_paid?: number | null
          updated_at?: string | null
        }
        Update: {
          agree_to_terms?: boolean | null
          created_at?: string | null
          customer_id?: string | null
          event_id?: string | null
          payment_status?: Database["public"]["Enums"]["payment_status"] | null
          primary_attendee_id?: string | null
          registration_data?: Json[] | null
          registration_date?: string | null
          registration_id?: string
          registration_type?:
            | Database["public"]["Enums"]["registration_type"]
            | null
          status?: string | null
          stripe_payment_intent_id?: string | null
          total_amount_paid?: number | null
          total_price_paid?: number | null
          updated_at?: string | null
        }
        Relationships: []
      }
// ...
      tickets: {
        Row: {
          attendee_id: string | null
          created_at: string | null
          event_id: string | null
          id: string
          is_partner_ticket: boolean | null
          registration_id: string | null
          ticket_price: number | null
          ticket_status: string | null
          ticket_type_id: string | null
          updated_at: string | null
        }
        Insert: {
          attendee_id?: string | null
          created_at?: string | null
          event_id?: string | null
          id: string
          is_partner_ticket?: boolean | null
          registration_id?: string | null
          ticket_price?: number | null
          ticket_status?: string | null
          ticket_type_id?: string | null
          updated_at?: string | null
        }
        Update: {
          attendee_id?: string | null
          created_at?: string | null
          event_id?: string | null
          id?: string
          is_partner_ticket?: boolean | null
          registration_id?: string | null
          ticket_price?: number | null
          ticket_status?: string | null
          ticket_type_id?: string | null
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "tickets_registration_id_fkey"
            columns: ["registration_id"]
            isOneToOne: false
            referencedRelation: "registrations"
            referencedColumns: ["registration_id"]
          },
        ]
      }
// ...
    }
// ...
    Enums: {
      attendee_contact_preference:
        | "directly"
        | "primaryattendee"
        | "mason"
        | "guest"
        | "providelater"
      attendee_type: "mason" | "guest" | "ladypartner" | "guestpartner"
      billing_reason:
        | "subscription_cycle"
        | "subscription_create"
        | "subscription_update"
        | "subscription_threshold"
        | "manual"
        | "upcoming"
        | "quote_accept"
      billing_scheme: "per_unit" | "tiered"
      collection_method: "charge_automatically" | "send_invoice"
      invoice_status: "draft" | "open" | "paid" | "void" | "uncollectible"
      organisation_type:
        | "lodge"
        | "grandlodge"
        | "masonicorder"
        | "company"
        | "other"
      payment_status:
        | "pending"
        | "completed"
        | "failed"
        | "refunded"
        | "partially_refunded"
        | "cancelled"
        | "expired"
      price_type: "one_time" | "recurring"
      quote_status: "draft" | "open" | "accepted" | "canceled" | "expired"
      registration_type:
        | "individuals"
        | "groups"
        | "officials"
        | "lodge"
        | "delegation"
      stripe_order_status: "pending" | "completed" | "canceled"
      stripe_subscription_status:
        | "not_started"
        | "incomplete"
        | "incomplete_expired"
        | "trialing"
        | "active"
        | "past_due"
        | "canceled"
        | "unpaid"
        | "paused"
      tax_behavior: "inclusive" | "exclusive" | "unspecified"
      URL: "url"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof Database },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never